+++ sdk/healthinsights/health-insights-radiologyinsights-rest/CHANGELOG.md
@@ -15,6 +15,7 @@
 - Client Changes
     - POST call replaced with PUT (this change is taken care of automatically by the SDK)
     - Adding TokenCredential next to KeyCredential
+    - Credential through InteractiveBrowserCredential (browser) and DefaultAzureCredential (node)
 
 - Request changes:
     - Renamed 'createdDateTime' into 'createdAt'